% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractUT.R
\name{extractUT}
\alias{extractUT}
\title{extractUT}
\usage{
extractUT(Y, dims, includeDiagonal = F)
}
\arguments{
\item{Y}{an array of dims x dims x N SPD matrices.}

\item{dims}{the dimension of the SPD matrices.}

\item{includeDiagonal}{if T, will return the diagonal elements as well as the upper triangular elements.}
}
\description{
extract the upper triangular elements of an array of SPDs, return data.frame, each row belongs to one observation
}
