% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimNoiseSPD.R
\name{genSPD_SNR}
\alias{genSPD_SNR}
\title{genSPD_SNR}
\usage{
genSPD_SNR(d, X, C, scale = 1, SNR = 1, bp = NULL, maxDist = 1.25,
  minDist = 0.75, maximumSPDValue = NULL)
}
\arguments{
\item{d}{Dimension of SPD matrices to generate.}

\item{X}{A x by n matrix, x rows indicate an observed variable, n columns an observed individual.}

\item{C}{A c by n matrix, c rows indicate an observed variable, n columns an observed individual.}

\item{scale}{scale factor to scale X and C (reduces computational issues from overly large SPD matrices).}

\item{SNR}{Signal to Noise ratio to be used in generating the noise. SNR must be larger 0.}

\item{bp}{Base point on SPD manifold (default is Identity).}

\item{maxDist}{maximum distance from bp for coefficient vectors.}

\item{minDist}{minimum distance from bp for coefficient vectors.}

\item{maximumSPDValue}{If not NULL, limits the maximum value that any entry in the SPD matrix can obtain.}
}
\description{
Generates SPD matrices with noise from covariates X, confound C, with signal to noise ratio SNR.
}
