% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimSPDData.R
\name{genSPDdata}
\alias{genSPDdata}
\title{genSPDdata}
\usage{
genSPDdata(N = 500, dims = 5, maxDist = 1, SNR = 1,
  includeDiagonal = F, beta = NULL)
}
\arguments{
\item{N}{number of observations (N is the number of SPD matrices generated)}

\item{dims}{dimension of SPD matrices to generate}

\item{maxDist}{maximum distance on the SPD manifold for the generated SPD matrices (measured from the identity matrix)}

\item{SNR}{Signal to Noise ratio (noise has standard deviation = 1/SNR)}

\item{includeDiagonal}{If T, will include the diagonal entries of the SPD matrix. Otherwise will only include the uppder diagonal elements.}

\item{beta}{vector of coefficients for the generated covariates. Default is rep(1, times=dims). If beta is too short, will be right padded with zeros. If beta is too long, will be right-truncated.}
}
\description{
Generate random SPD data with covariates (1 covariate per SPD upper diagonal entry). Default covariate coefficient is beta_j=1. Control the signal to noise ratio via SNR. Noise is gaussian and uncorrelated. A dataframe containing N rows, and 1 column for each matrix upper diagonal entry, and 1 column for each corresponding covariate. Generates Y_dimsxdims, y1...y_(n_y), and x1...x_(n_y). Returns a data.frame with 2*n_y columns, and N rows. n_y is dims*(dims/2+1/2), or dims*(dims/2-1/2) depending on whether the main diagonal is included. The y subsript counts along rows (left to right) first, then moves down one row at a time (eg for dims=3 when the main diagonal is excluded, y1 = Y[1,2], y2=Y[1,3], y3=Y[2,3]).
}
