% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimConfoundData.R
\name{SimConfoundData}
\alias{SimConfoundData}
\title{SimConfoundData}
\usage{
SimConfoundData(n, D, ...)
}
\arguments{
\item{n}{the sample size to use}

\item{D}{the distribution function to draw from (eg D=rbinom), must take n (sample size) as first parameter}

\item{...}{parameters to pass to the distribution function (eg p=0.25 for D=rbinom)}
}
\description{
SimConfoundData simulates a vector of confounds from a given distribution
}
\examples{
C1 <- SimConfoundData(n=10, D=rbinom, size=1, prob=0.25)
}
