% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimConfoundData.R
\name{SimSNPData}
\alias{SimSNPData}
\title{SimSNPData}
\usage{
SimSNPData(n, d, p = NULL)
}
\arguments{
\item{n}{the sample size}

\item{d}{the maximum number of SNP differences (eg if d=3, there will be four SNP categoris: 0, 1, 2, or 3 SNP differences)}

\item{p}{probability weights for each number of SNP differences}
}
\description{
SimSNPData simulates SNP differences. Wrapper for SimConfoundData
}
\examples{
C1 <- SimSNPData(n=20, d=3, p=c(7,2,1))
}
