% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimSPDData.R
\name{genSPDdata}
\alias{genSPDdata}
\title{genSPDdata}
\usage{
genSPDdata(N = 500, dims = 5, maxDist = 1, SNR = 1,
  includeDiagonal = F, beta = NULL, C = NULL, P = NULL)
}
\arguments{
\item{N}{number of observations (N is the number of SPD matrices generated)}

\item{dims}{dimension of SPD matrices to generate}

\item{maxDist}{maximum distance on the SPD manifold for the generated SPD matrices (measured from the identity matrix)}

\item{SNR}{Signal to Noise ratio (noise has standard deviation = 1/SNR)}

\item{includeDiagonal}{If T, will include the diagonal entries of the SPD matrix. Otherwise will only include the uppder diagonal elements.}

\item{beta}{If C is NULL, this will be used to generate data under a element-wise multivariate normal model (y1...yp ~ X*Beta), NOT a Riemann Manifold model. Vector of coefficients for the generated covariates. Default is rep(1, times=dims). If beta is too short, will be right padded with zeros. If beta is too long, will be right-truncated.}

\item{C}{If C is not NULL, this will be used to generate data under a Riemann Manifold model (Y ~ X*V), NOT a multivariate normal model. A kxN array of confounds (each row is an N-vector corresponding to a set of observations for one confound, each column corresponds to a k-vector of confound observations for the ith individual)}

\item{P}{If C is not NULL, P is the base point on the manifold from which V displaces (Y ~ exp_m(P,CV)), default is P=Identity matrix (dimension dims.}
}
\value{
If C=NULL (no confounds), returns a data.frame with 2*n_y columns, and N rows. n_y is dims*(dims/2+1/2), or dims*(dims/2-1/2) depending on whether the main diagonal is included. Otherwise, returns a list with elements Y, X, V, P, and yx, where Y is the list of observed SPD matrices, X is the data.frame of k confounds (Nxk), V is the list of k SPD covariate matrices, P is the base point on the manifold, and yx is the data.frame of upper diagonal elements of Y (with each row corresponding to one observation Y[[i]]).
}
\description{
Generate random SPD data with covariates (1 covariate per SPD upper diagonal entry). Default covariate coefficient is beta_j=1. Control the signal to noise ratio via SNR. Noise is gaussian and uncorrelated. A dataframe containing N rows, and 1 column for each matrix upper diagonal entry, and 1 column for each corresponding covariate. Generates Y_dimsxdims, y1...y_(n_y), and x1...x_(n_y). The y subsript counts along rows (left to right) first, then moves down one row at a time (eg for dims=3 when the main diagonal is excluded, y1 = Y[1,2], y2=Y[1,3], y3=Y[2,3]).
}
\examples{
set.seed(1234)
g <- genSPDdata(N = 3, dims = 3, SNR = 2, C=matrix(c(1,2,1),ncol=3))
g$Y[[1]]
MGLMRiem::expmap_spd(g$P,g$X[1]*g$V)
}
